fixed more warnings by casting a bunch of stuff

+++ src/man/vision/Ball.cpp
@@ -144,7 +144,7 @@ void Ball::preScreenBlobsBasedOnSizeAndColor() {
 	// pre-screen blobs that don't meet our criteria
 	for (int i = 0; i < blobs->number(); i++) {
         int ar = blobs->get(i).getArea();
-        float perc = blobs->get(i).getPixels() / static_cast<float>(ar);
+        float perc = static_cast<float>(blobs->get(i).getPixels()) / static_cast<float>(ar);
         int w = blobs->get(i).width();
         int h = blobs->get(i).height();
         int diam = max(w, h);
